add : PermissionViewModel 추가 및 권한 관련 함수 분리, 적용 중 ...
