+++ app/src/main/java/com/rocket/cosmic_detox/presentation/viewmodel/PermissionViewModel.kt
@@ -1,0 +1,29 @@
+package com.rocket.cosmic_detox.presentation.viewmodel
+
+import android.app.AppOpsManager
+import android.content.Context
+import android.os.Build
+import android.provider.Settings
+import androidx.lifecycle.ViewModel
+import dagger.hilt.android.lifecycle.HiltViewModel
+
+@HiltViewModel
+class PermissionViewModel : ViewModel() {
+    fun isUsageStatsPermissionGranted(context: Context): Boolean {
+        val appOps = context.getSystemService(Context.APP_OPS_SERVICE) as AppOpsManager
+        val mode = appOps.checkOpNoThrow(
+            AppOpsManager.OPSTR_GET_USAGE_STATS,
+            android.os.Process.myUid(),
+            context.packageName
+        )
+        return mode == AppOpsManager.MODE_ALLOWED
+    }
+
+    fun isOverlayPermissionGranted(context: Context): Boolean {
+        return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+            Settings.canDrawOverlays(context)
+        } else {
+            true
+        }
+    }
+}
